requirements.txt
allure-pytest==2.13.2
allure-python-commons==2.13.2
attrs==23.1.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.2.0
colorama==0.4.6
exceptiongroup==1.1.2
Faker==19.1.0
h11==0.14.0
idna==3.4
iniconfig==2.0.0
mypy==1.4.1
mypy-extensions==1.0.0
outcome==1.2.0
packaging==23.1
pluggy==1.2.0
pycparser==2.21
PySocks==1.7.1
pytest==7.4.0
pytest-ordering==0.6
python-dateutil==2.8.2
python-dotenv==1.0.0
requests==2.31.0
selenium==4.10.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
tomli==2.0.1
tqdm==4.65.0
trio==0.22.2
trio-websocket==0.10.3
typing_extensions==4.7.1
urllib3==2.0.3
webdriver-manager==3.8.6
wsproto==1.2.0
